employees-api/pyproject.toml
[tool.poetry]
name = "employees-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
faker = "^37.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
